+++ src/Materials.java
@@ -9,6 +9,7 @@ import javafx.scene.control.Button;
 import javafx.scene.control.MenuItem;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
+import javafx.scene.effect.DropShadow;
 import javafx.scene.input.*;
 import javafx.scene.layout.*;
 import javafx.scene.paint.Color;
@@ -91,7 +92,6 @@ class All_materials {
 class Center_scroller_pane{
     private ScrollPane mat_pane = new ScrollPane();
     private VBox mat_vbox = new VBox();
-    private VBox bkmark_vbox = new VBox();
     private ArrayList mat_list = new ArrayList<>();
     private ArrayList bkmark_list = new ArrayList();
     private ContextMenu menu;
@@ -108,6 +108,7 @@ class Center_scroller_pane{
         MenuItem itemAddMat = new MenuItem("添加资料");
         MenuItem itemAddBkm = new MenuItem("添加书签");
         menu.getItems().addAll(itemAddMat,itemAddBkm);
+        mat_pane.setStyle("-fx-background-color: #203A97");
         mat_pane.setOnMouseClicked(new EventHandler<MouseEvent>() {
             public void handle(MouseEvent event) {
                 if (event.getButton().equals(MouseButton.SECONDARY)) {
@@ -132,10 +133,11 @@ class Center_scroller_pane{
         }catch(NullPointerException exc){};
         mat_vbox.setSpacing(StaticValue.stageHeight / 20);
         mat_vbox.setPadding(new Insets(StaticValue.stageHeight/10,0,0,StaticValue.stageWidth/3));
-        Text title_mat = new Text("资料");
+        Text title_mat = new Text("本地资料");
         title_mat.setFont(Font.font("华文行楷", FontWeight.BOLD, 30));
         title_mat.setFill(Color.rgb(245,202,42));
         mat_vbox.getChildren().add(title_mat);
+        mat_vbox.setStyle("-fx-background-color: #203A97");
         Button buttons[] = new Button[mat_list.size()];
 
         for(i = 0;i < mat_list.size();i++){
@@ -164,6 +166,9 @@ class Center_scroller_pane{
             VBox.setMargin(options[i], new Insets(0, 0, 0, 8)); //为每个节点设置外边距
             mat_vbox.getChildren().add(options[i]);
         }
+        mat_vbox.setStyle("-fx-background-color: #203A97");
+        mat_pane.setFitToHeight(true);
+        mat_pane.setFitToWidth(true);
         mat_pane.setContent(mat_vbox);
     }
     /**
@@ -196,6 +201,16 @@ class Mat_button{
         MenuItem itemEdit = new MenuItem("编辑");
         menu.getItems().addAll(itemDel,itemEdit);
         button.setFont(Font.font("宋体",18));
+        //当鼠标进入按钮时添加阴影特效
+        DropShadow shadow = new DropShadow();
+        shadow.setColor(Color.rgb(67,173,217));
+        button.addEventHandler(MouseEvent.MOUSE_ENTERED, (MouseEvent e) -> {
+            button.setEffect(shadow);
+        });
+        //当鼠标离开按钮时移除阴影效果
+        button.addEventHandler(MouseEvent.MOUSE_EXITED, (MouseEvent e) -> {
+            button.setEffect(null);
+        });
         button.setOnMouseClicked(new EventHandler<MouseEvent>() {
             public void handle(MouseEvent event) {
                 if(event.getButton().equals(MouseButton.SECONDARY)) {
@@ -215,6 +230,7 @@ class Mat_button{
                 }
             }
         });
+        button.setStyle(StaticValue.buttonStyle2 + "-fx-font-size: 18");
         itemDel.setOnAction(e -> {
             new Material_alert("Delete",button.getText(),course,0);
         });
@@ -338,8 +354,11 @@ class Material_alert {
     Material_alert(String s,String course){
         if(s.equals("Add")){
             Text text_name = new Text("请输入书签名称：");
+            text_name.setFill(Color.rgb(245,202,42));
             Text text_content = new Text("请输入网址链接：");
+            text_content.setFill(Color.rgb(245,202,42));
             Text hint = new Text("(建议该网址从浏览器复制)");
+            hint.setFill(Color.rgb(245,202,42));
             text_name.setFont(Font.font("宋体",18));
             text_content.setFont(Font.font("宋体",18));
             TextField textField_name = new TextField();
@@ -348,6 +367,9 @@ class Material_alert {
             Button no = new Button("取消");
             yes.setFont(Font.font("Microsoft YaHei", 18));
             no.setFont(Font.font("Microsoft YaHei", 18));
+            String buttonStyle1 = StaticValue.buttonStyle1 + "-fx-font-size: 18";
+            yes.setStyle(buttonStyle1);
+            no.setStyle(buttonStyle1);
             yes.setPadding(new Insets(10));
             no.setPadding(new Insets(10));
             HBox hBox1 = new HBox(text_name,textField_name);
@@ -359,7 +381,7 @@ class Material_alert {
             hBox2.setSpacing(StaticValue.stageHeight / 24);
             BorderPane borderPane = new BorderPane();
             borderPane.setCenter(vBox);
-
+            borderPane.setStyle("-fx-background-color: #203A97");
             borderPane.setBottom(hBox2);
             borderPane.setMargin(vBox,new Insets(30));
             Stage alert = new Stage();
@@ -367,13 +389,32 @@ class Material_alert {
             alert.setScene(new Scene(borderPane,StaticValue.stageWidth * 2 / 5,StaticValue.stageHeight * 2 / 5));
             alert.setTitle("提示信息");
             alert.show();
-
+            yes.setOnMouseMoved(e -> {
+                yes.setStyle(StaticValue.buttonStyle2 + "-fx-font-size: 18");
+            });
+            yes.setOnMouseExited(e -> {
+                yes.setStyle(StaticValue.buttonStyle1 + "-fx-font-size: 18");
+            });
+            no.setOnMouseMoved(e -> {
+                no.setStyle(StaticValue.buttonStyle2 + "-fx-font-size: 18");
+            });
+            no.setOnMouseExited(e -> {
+                no.setStyle(StaticValue.buttonStyle1 + "-fx-font-size: 18");
+            });
             no.setOnAction(e -> {
                 alert.close();
             });
 
             yes.setOnAction(e -> {
 //              书签链接文件后缀必须为.txt
+                if(textField_name.getText().equals("")){
+                    new LoginAlert("请输入正确的书签名！");
+                    return;
+                }
+                if(textField_content.getText().equals("")){
+                    new LoginAlert("请输入正确的网址！");
+                    return;
+                }
                 String new_bkm = textField_name.getText() + ".txt";
                 File f = new File("./data/" + StaticValue.userName + "/资源/" + course + "/书签链接/" + new_bkm);
                 if(f.exists()){
@@ -416,6 +457,8 @@ class Material_alert {
             text1 = new Text("可将本地文件拖拽至此");
             text.setFont(Font.font("宋体",18));
             text1.setFont(Font.font("宋体",16));
+            text.setFill(Color.rgb(245,202,42));
+            text1.setFill(Color.rgb(245,202,42));
             TextField textField = new TextField();
             Button yes = new Button("确定");
             Button no = new Button("取消");
@@ -423,6 +466,9 @@ class Material_alert {
             no.setFont(Font.font("Microsoft YaHei", 18));
             yes.setPadding(new Insets(10));
             no.setPadding(new Insets(10));
+            String buttonStyle1 = StaticValue.buttonStyle1 + "-fx-font-size: 18";
+            yes.setStyle(buttonStyle1);
+            no.setStyle(buttonStyle1);
             HBox hBox1 = new HBox(text,textField);
             VBox vBox = new VBox(hBox1,text1);
             vBox.setSpacing(StaticValue.stageHeight / 30);
@@ -433,7 +479,7 @@ class Material_alert {
             borderPane.setCenter(vBox);
             borderPane.setBottom(hBox2);
             borderPane.setMargin(vBox,new Insets(30));
-
+            borderPane.setStyle("-fx-background-color: #203A97");
             Stage alert = new Stage();
             alert.initModality(Modality.APPLICATION_MODAL);
             Scene mat_scene = new Scene(borderPane,StaticValue.stageWidth * 2 / 5,StaticValue.stageHeight * 2 / 5);
@@ -479,8 +525,20 @@ class Material_alert {
             });
             alert.setScene(mat_scene);
             alert.setTitle("提示信息");
+            alert.setHeight(200);
             alert.show();
-
+            yes.setOnMouseMoved(e -> {
+                yes.setStyle(StaticValue.buttonStyle2 + "-fx-font-size: 18");
+            });
+            yes.setOnMouseExited(e -> {
+                yes.setStyle(StaticValue.buttonStyle1 + "-fx-font-size: 18");
+            });
+            no.setOnMouseMoved(e -> {
+                no.setStyle(StaticValue.buttonStyle2 + "-fx-font-size: 18");
+            });
+            no.setOnMouseExited(e -> {
+                no.setStyle(StaticValue.buttonStyle1 + "-fx-font-size: 18");
+            });
             no.setOnAction(e -> {
                 alert.close();
             });
@@ -488,6 +546,10 @@ class Material_alert {
             yes.setOnAction(e -> {
                 if(mat_or_bkm == 0){
                     String new_mat = textField.getText();
+                    if(new_mat.equals("")){
+                        new LoginAlert("请输入正确的文件名！");
+                        return;
+                    }
                     File f = new File("./data/" + StaticValue.userName + "/资源/" + course + "/" + new_mat);
                     if(f.exists()) {
                         alert.close();
@@ -504,53 +566,32 @@ class Material_alert {
                     alert.close();
                     new Material_alert("SuccessAdd",course,0);
                 }
-                else{
-//                    书签链接文件后缀必须为.txt
-                    String new_bkm = textField.getText() + ".txt";
-                    File f = new File("./data/" + StaticValue.userName + "/资源/" + course + "/书签链接/" + new_bkm);
-                    try {
-                        f.createNewFile();
-                    } catch (IOException ex) {
-                        alert.close();
-                        new Material_alert("Exist",course,1);
-                        return;
-                    }
-                    alert.close();
-                    new Material_alert("SuccessAdd",course,1);
-                }
 
             });
         }
 //        提示信息——编辑成功
         else if(s.equals("SuccessEdit")) {
-            Alert alert = new Alert(Alert.AlertType.INFORMATION);
-            alert.setTitle("提示信息");
-            alert.setHeaderText("");
-            alert.setContentText("更改成功！");
-            alert.show();
+//            Alert alert = new Alert(Alert.AlertType.INFORMATION);
+//            alert.setTitle("提示信息");
+//            alert.setHeaderText("");
+//            alert.setContentText("更改成功！");
+//            alert.show();
             new Materials(Materials.material_pane,course);
+            new LoginAlert("编辑成功！");
         }
 //        提示信息——添加成功
         else if(s.equals("SuccessAdd")){
-            Alert alert = new Alert(Alert.AlertType.INFORMATION);
-            alert.setTitle("提示信息");
-            alert.setHeaderText("");
-            alert.setContentText("添加成功！");
-            alert.show();
             new Materials(Materials.material_pane,course);
+            new LoginAlert("添加成功！");
         }
 //        提示信息——已存在
         else if(s.equals("Exist")) {
-            Alert alert = new Alert(Alert.AlertType.INFORMATION);
-            alert.setTitle("提示信息");
-            alert.setHeaderText("");
             if(mat_or_bkm == 0){
-                alert.setContentText("该文件已存在！");
+                new LoginAlert("该文件已存在！");
             }
             else {
-                alert.setContentText("该书签已存在！");
+                new LoginAlert("该书签已存在！");
             }
-            alert.show();
         }
     }
     /**
@@ -598,6 +639,7 @@ class Material_alert {
                             ex.printStackTrace();
                         }
                         new Materials(Materials.material_pane,course);
+                        new LoginAlert("删除成功！");
                     }
                     else{
                         try {
@@ -607,6 +649,7 @@ class Material_alert {
                             ex.printStackTrace();
                         }
                         new Materials(Materials.material_pane,course);
+                        new LoginAlert("删除成功！");
                     }
 
                 } else {
@@ -625,6 +668,7 @@ class Material_alert {
                 text = new Text("请输入新书签名称：");
             }
             text.setFont(Font.font("宋体",18));
+            text.setFill(Color.rgb(245,202,42));
             TextField textField = new TextField();
             Button yes = new Button("确定");
             Button no = new Button("取消");
@@ -632,6 +676,9 @@ class Material_alert {
             no.setFont(Font.font("Microsoft YaHei", 18));
             yes.setPadding(new Insets(10));
             no.setPadding(new Insets(10));
+            String buttonStyle1 = StaticValue.buttonStyle1 + "-fx-font-size: 18";
+            yes.setStyle(buttonStyle1);
+            no.setStyle(buttonStyle1);
             HBox hBox1 = new HBox(text,textField);
             HBox hBox2 = new HBox(yes,no);
             hBox2.setAlignment(Pos.CENTER);
@@ -640,12 +687,25 @@ class Material_alert {
             borderPane.setCenter(hBox1);
             borderPane.setBottom(hBox2);
             borderPane.setMargin(hBox1,new Insets(30));
+            borderPane.setStyle("-fx-background-color: #203A97");
             Stage alert = new Stage();
             alert.initModality(Modality.APPLICATION_MODAL);
             alert.setScene(new Scene(borderPane,StaticValue.stageWidth * 2 / 5,StaticValue.stageHeight * 2 / 5));
             alert.setTitle("提示信息");
+            alert.setHeight(180);
             alert.show();
-
+            yes.setOnMouseMoved(e -> {
+                yes.setStyle(StaticValue.buttonStyle2 + "-fx-font-size: 18");
+            });
+            yes.setOnMouseExited(e -> {
+                yes.setStyle(StaticValue.buttonStyle1 + "-fx-font-size: 18");
+            });
+            no.setOnMouseMoved(e -> {
+                no.setStyle(StaticValue.buttonStyle2 + "-fx-font-size: 18");
+            });
+            no.setOnMouseExited(e -> {
+                no.setStyle(StaticValue.buttonStyle1 + "-fx-font-size: 18");
+            });
             no.setOnAction(e -> {
                 alert.close();
             });
@@ -658,6 +718,11 @@ class Material_alert {
 
                     File f1 = new File("./data/" + StaticValue.userName + "/资源/" + course + "/" + name);
                     File f2 = new File("./data/" + StaticValue.userName + "/资源/" + course + "/" + new_mat);
+                    if(f2.exists()) {
+                        alert.close();
+                        new LoginAlert("该文件已存在");
+                        return;
+                    }
                     alert.close();
                     f1.renameTo(f2);
                     new Material_alert("SuccessEdit",course,0);
@@ -668,6 +733,11 @@ class Material_alert {
                     String new_bkm = textField.getText()+".txt";
                     File f1 = new File("./data/" + StaticValue.userName + "/资源/" + course + "/书签链接/" + name);
                     File f2 = new File("./data/" + StaticValue.userName + "/资源/" + course + "/书签链接/" + new_bkm);
+                    if(f2.exists()) {
+                        alert.close();
+                        new LoginAlert("该书签已存在");
+                        return;
+                    }
                     alert.close();
                     f1.renameTo(f2);
                     new Material_alert("SuccessEdit",course,1);
